Move cloud config layer tests
Keeps cloud_config_layers.rs focused on implementation by moving its unit tests into cloud_config_layers_tests.rs, following the requirements_layers organization feedback. Adds a module overview comment for the cloud fragment-to-layer conversion path; behavior is unchanged.
